package/python3: set *_CONFIG_SCRIPTS variables
This change defines the pyhton3 ABI suffixes (c.f. PEP3149 [1]), and sets
the *_CONFIG_SCRIPTS variables for both host and target packages because,
for the host-python3 package, the python-config script installation depends
on the python3 package selection.

Note that *-config scripts are automatically removed from the target tree
by the infrastructure. So, this change also simplify the
PYTHON3_REMOVE_USELESS_FILES hook.

[1] https://www.python.org/dev/peps/pep-3149/

Signed-off-by: Samuel Martin <s.martin49@gmail.com>