package/python: set *_CONFIG_SCRIPTS variables
This change sets the *_CONFIG_SCRIPTS variables for both host and target
packages because, for the host-python package, the python-config script
installation depends on the python3 package selection.

Note that *-config scripts are automatically removed from the target tree
by the infrastructure. So, this change also simplify the
PYTHON_REMOVE_USELESS_FILES hook.

Signed-off-by: Samuel Martin <s.martin49@gmail.com>